refactor: align stream-tool demos with the shared chat + review UI
